add INSTANCE of mapper, mini-refactoring

+++ Forum/src/main/java/serverDb/service/ServiceService.java
@@ -10,7 +10,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class ServiceService {
@@ -34,19 +33,19 @@ public class ServiceService {
 
         String sql;
         sql = "SELECT count(id) from Post";
-        final int post = (int) jdbcTemplate.queryForObject(sql, Integer.class);
+        final int post = jdbcTemplate.queryForObject(sql, Integer.class);
         responseBody.put("post", post);
 
         sql = "SELECT count(slug) from Thread";
-        final int thread = (int) jdbcTemplate.queryForObject(sql, Integer.class);
+        final int thread = jdbcTemplate.queryForObject(sql, Integer.class);
         responseBody.put("thread", thread);
 
         sql = "SELECT count(slug) from Forum";
-        final int forum = (int) jdbcTemplate.queryForObject(sql, Integer.class);
+        final int forum = jdbcTemplate.queryForObject(sql, Integer.class);
         responseBody.put("forum", forum);
 
         sql = "SELECT count(nickname) from FUser";
-        final int user = (int) jdbcTemplate.queryForObject(sql, Integer.class);
+        final int user = jdbcTemplate.queryForObject(sql, Integer.class);
         responseBody.put("user", user);
 
         return new ResponseEntity(responseBody, HttpStatus.OK);
